requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
ai-edge-litert==2.1.4
numpy==2.4.4
SQLAlchemy==2.0.49
python-multipart==0.0.26
python-dotenv==1.2.2
pillow==12.2.0
anthropic==0.94.0
google-genai==1.72.0
requests==2.33.1
